requirements.txt
iisignature=0.24
jupyterlab=3.4.6
matplotlib=3.5.2
numpy=1.21.5
pandas=1.3.5
pytorch=1.10.2
sacred=0.8.2
scikit-learn=1.0.2
scipy=1.7.3
seaborn=0.11.2
stochastic=0.6.0
torchcde=0.2.5
torchdiffeq=0.2.3
torchsde=0.2.5